fix(cmake)!: automatic component and target discovery fixed
This change is breaking, as it changes the internal (though *not yet
documented*) properties used to track discovered components and targets.
We also properly import targets for packages *without* components. That
code was never implemented. The reason is simply "I Forgor 💀", however
there is very little to be done on the matter.

It also now makes it so that `ixm::package::check` *does* perform the
call to `ixm::package::import` as documented.

Furthermore, we now have a default value for `NAMES` for `ixm::find::*`
commands used for simple packages that only have a single executable or
name.

We also now automatically set the `DESCRIPTION` parameter and it is no
longer a required one for simple calls.

This PR also fixes the `find_package(clang-tidy)` module, as it was
never finding any targets.

Lastly, we organized the `internal/find.cmake` file slightly, though it
is still quite a large file.

Signed-off-by: Izzy Muerte <63051+bruxisma@users.noreply.github.com>