feat(cmake): add `target_compiler_launcher` command (#14)
This command provides a new and simplified way to set a target's
compiler launcher without having to worry about various minutiae for the
Xcode generator, the MSVC compiler and ABI (and by extension the Visual
Studio generator), and will continue to work even if precompiled headers
are used.

There are a few potential landmines specifically regarding the
`MSVC_DEBUG_INFORMATION_FORMAT` generator expression failing, as well as
instances where a caching tool was not used for the Xcode launcher.

Eventually, we'll figure out a way to handle these edge cases, but for
now we've already got this working, more or less.

---------

Signed-off-by: Izzy Muerte <63051+bruxisma@users.noreply.github.com>