fix(cmake): properly disable unnecessary targets
One of the features IXM is purportedly supposed to do is disable the
various CDash targets. Due to a cleanup prior to the repository being
made available, the actual properties to set this were in the wrong
file.

This PR rectifies this, and also places predefined targets into a CMake
folder.

Signed-off-by: Izzy Muerte <63051+bruxisma@users.noreply.github.com>