feat(cmake): add `ixm_fallback` function (#37)
This brings a new function, `ixm_fallback`, which replaces an internal
command `🈯::ixm::default`, and is now available for consumers. This
command is used to assign values to variables that have not yet been
defined. This is most useful when `cmake_parse_arguments` has been
called and an optional value was not passed in to the user.

Behavior regarding it's interaction with CMP0174 has been documented as
well.

Signed-off-by: Izzy Muerte <63051+bruxisma@users.noreply.github.com>