feat(cmake): add `ixm_property` command to the public runtime (#40)
This introduces `ixm_property`, a command that was originally
implemented as `ixm_target_property` in #21. However, as that PR is
broken apart and into separate pieces I've been re-evaluating how each
component should be named or how it works. Not much has changed with
this command's implementation however it's name required an update as
only target properties can be used with this command anyhow. There are
possibly more useful additions that can be added for the final generator
expression that is generated, however I'm keeping it as simple as it can
be for now given its use-cases.

Signed-off-by: Izzy Muerte <63051+bruxisma@users.noreply.github.com>