chore(config): reduce commitlint and renovate details (#42)
It's easier to just rely on the `type-enum` and change the scope in the
title of PRs from myself or others than to deal with it as a hard error
requirement. This also means we'll go back to `chore(deps)` but I'm fine
with that. We don't want those to show up in our changelog once we start
using `git-cliff` anyhow.

Signed-off-by: Izzy Muerte <63051+bruxisma@users.noreply.github.com>