ci(actions): build documentation when a PR is opened (#43)
After the recent deployment failure, checking to make sure documentation
at least builds if any file under `docs/` is built is worth the time we
run the task for. Currently using the builtin `paths` argument, but
might switch to the skip-duplicate-actions workflow so that we can mark
this workflow as required.

Additionally, we're not doing anything with the output, we just want to
make sure there aren't any build errors from vitepress at this time. If
we move to starlight in the future, we'll most likely keep this same
approach/workflow.

---------

Signed-off-by: Izzy Muerte <63051+bruxisma@users.noreply.github.com>